updated the code based on the comment.
