--- src/components/landing/volunteerList.js
@@ -1,41 +1,0 @@
-import React, {Component} from "react";
-
-function volunteerlist () 
-{
-    return (
-      <React.Fragment>
-     
-    <table class="table table-hover">
-    <thead>
-      <tr>
-        <th scope="col">#</th>
-        <th scope="col">User Name</th>
-        <th scope="col">Joined Date</th>
-        <th scope="col">Level</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <th scope="row">1</th>
-        <td>Mark</td>
-        <td>9/29/2019</td>
-        <td>2</td>
-      </tr>
-      <tr>
-        <th scope="row">2</th>
-        <td>Jacob</td>
-        <td>8/17/2018</td>
-        <td>4</td>
-      </tr>
-      <tr>
-        <th scope="row">3</th>
-        <td>Larry the Bird</td>
-        <td>7/15/2016</td>
-        <td>5</td>
-      </tr>
-    </tbody>
-  </table>
-  </React.Fragment>
-    )
-}
-export default volunteerlist;
